Fixed bad optimization
Replaced a hinted, guarded set with a branchless set. The guard was preventing from setting a bool to true if it was already true, and hinted that it would already be true. Simply assigning true to an already true bool is likely less expensive than branching to avoid setting the value without change.